Handle watching common and node-server code from dev mode

+++ portal/vite.config.ts
@@ -1,7 +1,15 @@
 import { paraglide } from '@inlang/paraglide-sveltekit/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
-import { spawn } from 'child_process';
+import { spawn, type ChildProcessWithoutNullStreams } from 'child_process';
 import { stat, writeFile } from 'fs/promises';
+import {
+  createEmitAndSemanticDiagnosticsBuilderProgram,
+  createWatchCompilerHost,
+  createWatchProgram,
+  sys,
+  type BuilderProgram,
+  type Watch
+} from 'typescript';
 import { loadEnv, searchForWorkspaceRoot } from 'vite';
 import { defineConfig } from 'vitest/config';
 
@@ -54,30 +62,53 @@ export default defineConfig({
     },
     paraglide({ project: './project.inlang', outdir: './src/lib/paraglide' }),
     sveltekit(),
-    {
-      name: 'Run BullMQ Worker',
-      configureServer(server) {
-        if (server.config.mode === 'development') {
-          const compilingProcess = spawn('npx', ['tsc', '--project', 'tsconfig.dev.json'], {
-            cwd: 'node-server',
-            stdio: 'pipe',
-            shell: true
-          });
-          compilingProcess.once('close', () => {
-            const env = Object.assign(loadEnv(server.config.mode, process.cwd()), process.env);
-            const executingProcess = spawn('node', ['dev.js'], {
-              cwd: 'node-server',
+    (() => {
+      let executingProcess: ChildProcessWithoutNullStreams;
+      let running: boolean;
+      let watchProgram: Watch<BuilderProgram>;
+      return {
+        name: 'Run BullMQ Worker',
+        configureServer(server) {
+          if (server.config.mode === 'development') {
+            if (watchProgram) return;
+            spawn('tsc', ['-w'], {
+              cwd: 'common',
               stdio: 'pipe',
-              shell: true,
-              env
+              shell: true
             });
-            executingProcess.stderr.on('data', (dat) => process.stderr.write(dat));
-            executingProcess.stdout.on('data', (dat) => process.stdout.write(dat));
-            server.httpServer?.on('close', () => executingProcess.kill());
-          });
+            const env = Object.assign(loadEnv(server.config.mode, process.cwd()), process.env);
+            watchProgram = createWatchProgram(
+              createWatchCompilerHost(
+                'node-server/tsconfig.dev.json',
+                {},
+                sys,
+                createEmitAndSemanticDiagnosticsBuilderProgram,
+                () => {},
+                async () => {
+                  if (executingProcess) {
+                    if (running) return;
+                    running = true;
+                    const prom = new Promise((r) => executingProcess.once('close', r));
+                    executingProcess.kill();
+                    await prom;
+                  }
+                  executingProcess = spawn('node', ['dev.js'], {
+                    cwd: 'node-server',
+                    stdio: 'pipe',
+                    shell: true,
+                    env
+                  });
+                  executingProcess.stderr.on('data', (dat) => process.stderr.write(dat));
+                  executingProcess.stdout.on('data', (dat) => process.stdout.write(dat));
+                  server.httpServer?.on('close', () => executingProcess.kill());
+                  running = false;
+                }
+              )
+            );
+          }
         }
-      }
-    }
+      };
+    })()
   ],
   test: {
     include: ['src/**/*.{test,spec}.{js,ts}']
